Fixed quoteing bug.
If a token is wrapped in quotes (not just the term, the entire token) is
wrapped in double quotes, it should automatically be treated as a simple
token no matter what the internal format, e.g., '"foo: bar"' should
become a Token with operator=nil and term="foo: bar".

It would probably be best to be stingy with what characters are
acceptable in operators -- I like [a-zA-Z0-9] being it. We need to
consolidate the regular expressions for breaking scanning and breaking
up tokens anyway. (Different modules require different groupings, but we
can at least define the components in a single place.)