Consolidated date pasers, use AREL to build query
This change is not backwards-compatible with the old date parsers. The
constructor for DateParser now expects an AREL Attribute as an argument
instead of the table name and column name. An ActiveRecord Connection is
no longer required since the parser is no longer responsible for quoting
table and column names.